more problem solved

+++ Day_01 programme/SelfProblem_03.java
@@ -6,6 +6,6 @@ public class SelfProblem_03 {
         Scanner sc = new Scanner(System.in);
         int base = sc.nextInt();
         int exponent = sc.nextInt();
-        System.out.println(Math.pow(base, exponent));
+        System.out.println((int)Math.pow(base, exponent));
     }
 }
